Added Aria true/false to header.njk
Following comment is also in issues:

Found this code which I added to header.njk

Not sure if this will work. Inserting JS into HTML and I'm a little cloudy on everything the JS is doing.

Basically I copy/pasted aria code to add the property to c-header__nav-container js-nav-panel

<div class="c-header__nav-container js-nav-panel">
	{% include "components/tree-nav.njk" %}

	<script type="text/javascript">
		function toggle(id, id2) {
			var n = document.getElementById(id);
			if (n.style.display != 'none')
			         {
				n.style.display = 'none';
				document.getElementById(id2).setAttribute('aria-expanded', 'false');
			         }
			else
			        {
			n.style.display = '';
			document.getElementById(id2).setAttribute('aria-expanded', 'true');
				}
			}
	</script>
</div>