Added elements to new page.
I put everything in the `resouces.md`  I initially created a `resources.njk` file because I thought that's where the html should go but I wasn't sure how to reference the njk to get it to show up in the browser.  Also, weird thing.  For "Style Guides..." and "Examples" I had to create seperate divs for the uls in order to get them to work.  Not sure why, but they wouldn't work if they were in the same div as the h2s.