Scrub Hylia
I left a lot in the read me.  Mainly because there are some reference links and accredidation to the authors.  It's to the point where that readme could be deleted all together.  I did add a colophon to give credit to Hylia. Just let me know how you want to treat that readme.  Either changing Hylia to patternlab-website or deleting the readme will take me like 2 seconds, so I figured I'd ask.