Blog post index and templates
Set up so that new blog posts are added under posts and listed most recent add first.  However...

I'm getting an error back which I believe is causing the template builds to break.  The error is in dist relating to blog1.  The directory looks like it is going to the right place.  So I'm not sure what the issue is.