bug fix in updates #10
Just restarted my preview and it fired up fine.  This change in blog.njk (which doesn't look like a change to me) popped up after I re-booted. Looks like it was a glitch in the matrix.  Worth keeping open for testing again to maybe see if it's a recurring issue.  For now, it's working though.